requirements.txt
fastapi
uvicorn[standard]
streamlit
langchain
langchain-openai
pandas
python-dotenv
requests
